Remove circular ref
